ModelGenerator-pytorch/requirements.txt
pytorch
torchsummary
torchvision
pytorchviz

codecov
pytest
pytest-cov
matplotlib
scipy
pillow
tqdm